requirements.txt
# 核心依赖
websockets==12.0
sounddevice==0.4.6
numpy==1.24.3
scipy==1.14.1
python-dotenv==1.0.0

# 日志和调试
colorlog==6.8.0

# 系统监控
psutil==5.9.8

# 测试依赖（可选）
pytest==7.4.3
pytest-asyncio==0.21.1

# macOS 菜单栏应用
rumps==0.4.0
pyobjc==10.3
pyobjc-framework-AVFoundation==10.3
pyobjc-framework-CoreAudio==10.3
pyobjc-framework-CoreMedia==10.3
keyring==24.3.1
